+++ Projects/Domain/ProductDomain/Sources/Repository/ProductRepositoryImpl.swift
@@ -29,11 +29,11 @@ public struct ProductRepositoryImpl: ProductRepository {
         remoteProductDataSource.fetchProductList(page: page, text: text, tagIds: tagIds)
     }
     
-    public func createReview(id: Int,content: String, rating: Float, images: [String]) -> RxSwift.Single<BaseDomainInterface.BaseEntity> {
+    public func createReview(id: Int,content: String, rating: Int, images: [String]) -> RxSwift.Single<BaseDomainInterface.BaseEntity> {
         remoteProductDataSource.createReview(id: id,content: content, rating: rating, images: images)
     }
     
-    public func editReview(id: Int, reviewId: Int, content: String, rating: Float, images: [String]) -> RxSwift.Single<BaseDomainInterface.BaseEntity> {
+    public func editReview(id: Int, reviewId: Int, content: String, rating: Int, images: [String]) -> RxSwift.Single<BaseDomainInterface.BaseEntity> {
         
         remoteProductDataSource.editReview(id: id, reviewId: reviewId, content: content, rating: rating, images: images)
     }
