+++ Projects/Domain/ProductDomain/Sources/UseCase/CreateReviewUseCaseImpl.swift
@@ -20,7 +20,7 @@ struct CreateReviewUseCaseImpl: CreateReviewUseCase {
         self.productRepository = productRepository
     }
     
-    func execute(id: Int,content: String, rating: Float, images: [String]) -> Single<BaseEntity> {
+    func execute(id: Int,content: String, rating: Int, images: [String]) -> Single<BaseEntity> {
         productRepository.createReview(id: id,content: content, rating: rating, images: images)
     }
